# fix(gatsby-plugin-mdx): Use getNodesByType for plugin transformation