--- a/packages/gatsby-plugin-mdx/utils/get-source-plugins-as-remark-plugins.js
+++ b/packages/gatsby-plugin-mdx/utils/get-source-plugins-as-remark-plugins.js
@@ -13,7 +13,7 @@ module.exports = async function getSourcePluginsAsRemarkPlugins({
   gatsbyRemarkPlugins,
   mdxNode,
   getNode,
-  getNodes,
+  getNodesByType,
   reporter,
   cache,
   pathPrefix,
@@ -39,7 +39,7 @@ module.exports = async function getSourcePluginsAsRemarkPlugins({
       }
   }
 
-  fileNodes = getNodes().filter(n => n.internal.type === `File`)
+  fileNodes = getNodesByType(`File`)
 
   // return list of remarkPlugins
   const userPlugins = gatsbyRemarkPlugins